# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Axum Production API",
    "description": "Production-ready Axum API scaffold: session-backed auth, M2M API keys, RBAC, apps, organizations, owned storage with signed URLs, rate limiting, idempotency, metrics, cross-replica realtime, notifications, and an append-only audit log",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/change-password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Password changed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Keyset-paginated, and scoped to the calling user by the query itself.",
        "operationId": "list_notifications",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Limit number of items (default: 20, max: 100)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor token for next page",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "org_id",
            "in": "query",
            "description": "Restrict the feed to one organization. Membership is verified, so this\ncannot be used to read a tenant you do not belong to.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification page for the current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_Notification"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/notifications/read-all": {
      "patch": {
        "tags": [
          "Notifications"
        ],
        "operationId": "mark_all_notifications_read",
        "responses": {
          "200": {
            "description": "All marked as read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/notifications/{id}/read": {
      "patch": {
        "tags": [
          "Notifications"
        ],
        "operationId": "mark_notification_read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notification ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Marked as read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "list_users",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (1-indexed, default: 1)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Items per page (default: 20, max: 100)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated users (admin only)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_UserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin privileges required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "get_me",
        "responses": {
          "200": {
            "description": "Current authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "operationId": "delete_me",
        "responses": {
          "200": {
            "description": "Account deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "operationId": "update_me",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/users/me/password": {
      "patch": {
        "tags": [
          "Users"
        ],
        "operationId": "change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Password changed; other sessions revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect current password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorPayload": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {},
          "message": {
            "type": "string"
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "description": "Standardized Error Envelope",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorPayload"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponse_String": {
        "type": "object",
        "description": "Standardized Success Envelope",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "type": "string"
          },
          "meta": {},
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponse_UserResponse": {
        "type": "object",
        "description": "Standardized Success Envelope",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id",
              "name",
              "email",
              "email_verified",
              "role",
              "created_at"
            ],
            "properties": {
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "email": {
                "type": "string"
              },
              "email_verified": {
                "type": "boolean"
              },
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "image": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "name": {
                "type": "string"
              },
              "role": {
                "type": "string"
              }
            }
          },
          "meta": {},
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponse_Vec_Notification": {
        "type": "object",
        "description": "Standardized Success Envelope",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "user_id",
                "title",
                "body",
                "read",
                "created_at"
              ],
              "properties": {
                "body": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "data": {},
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "org_id": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid",
                  "description": "Set when the notification concerns an organization rather than the\nrecipient personally."
                },
                "read": {
                  "type": "boolean"
                },
                "title": {
                  "type": "string"
                },
                "user_id": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "meta": {},
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponse_Vec_UserResponse": {
        "type": "object",
        "description": "Standardized Success Envelope",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "name",
                "email",
                "email_verified",
                "role",
                "created_at"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "email": {
                  "type": "string"
                },
                "email_verified": {
                  "type": "boolean"
                },
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "image": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "name": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                }
              }
            }
          },
          "meta": {},
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ChangePasswordRequest": {
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "properties": {
          "current_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string"
          }
        }
      },
...
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "image": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}